requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
kafka-python>=2.0.2
pyspark==3.5.1
streamlit>=1.33.0
